+++ src/system/MovementSys.cpp
@@ -22,8 +22,8 @@ void MovementSys::Update_()
 	m_rReg.view<PlayerComp, PositionComp>().each([&](auto entity, auto& positionComp) {
 		{
 			auto& lastPositionComp = m_rReg.get_or_emplace<LastPositionComp>(entity);
-			int& xPos = positionComp.m_position.x;
-			int& yPos = positionComp.m_position.y;
+			float& xPos = positionComp.m_position.x;
+			float& yPos = positionComp.m_position.y;
 			lastPositionComp.m_lastPosition.x = xPos;
 			lastPositionComp.m_lastPosition.y = yPos;
 			using namespace sf;
