+++ src/util/ApplicationParameters.cpp
@@ -10,7 +10,7 @@ std::string const ApplicationParameters::k_windowName = "Emmoria";
 std::string const ApplicationParameters::k_dataPath = "data/";
 std::string const ApplicationParameters::k_spritePath = "sprite/";
 
-int const ApplicationParameters::k_playerMovementSpeed = 3;
+float const ApplicationParameters::k_playerMovementSpeed = 3;
 int const ApplicationParameters::k_widthUnits = 120;
 int const ApplicationParameters::k_heightUnits = 108;
 int const ApplicationParameters::k_debounceFrames = 10;
